Dev branch (#12)
* MOD: CheckPoint - RollBack
The following functionalities have been  completed. Namely
1. Coding all Functionalities
2. Tested all Functionalities
3. If Job/View is not present on the Interim. Delete the same on the Production.

This is the checkpoint before I go ahead and add the rich module for a more interactive console output.

* MOD: CheckPoint - RollBack
The following functionalities have been  completed. Namely
1. Coding all Functionalities
2. Tested all Functionalities
3. If Job/View is not present on the Interim. Delete the same on the Production.

This is the checkpoint before I go ahead and add the rich module for a more interactive console output.

* ADD: Creating a Bsae Module for Jenkins

---------

Co-authored-by: Joel_Thomas_Chacko <joel.chacko@tallysolutions.com>